implemented- i think correctly in this  way but the u(a*b) idk if == u(a)*u(b) problem is to be investigated and possibly fixed in the future

+++ piedPiper/implementacje/Hipki/HipekOcenaOgolnaProcessor.cs
@@ -12,7 +12,7 @@ namespace piedPiper.implementacje.Hipki
     {
         private float passingScore = 0;
 
-        public HipekOcenaOgolnaProcessor(int passingScore)
+        public HipekOcenaOgolnaProcessor(float passingScore)
         {
             this.passingScore = passingScore;
         }
